wip: add logo apk and transaction

+++ front-end/Takana/app/src/main/java/com/example/takana/presentation/money_account/MoneyAccountAddEditActivity.kt
@@ -57,7 +57,7 @@ class MoneyAccountAddEditActivity : AppCompatActivity() {
                         etAccountAmount.text.toString().toLong(),
                         accountTypeId,
                         user?.userId!!,
-                        etBankAccountNumber.text.toString().toInt()
+                        etBankAccountNumber.text.toString()
                     )
                     viewModelAddMoneyAccount()
                 }
